chore: batch-landing fixes (node-api scoped ptrs, header/ic_miss splits, fmt)
